+++ CHANGELOG.md
@@ -20,6 +20,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Tests for "The electrician apprentice".
 - Tests for "Sudoku validator".
 - Tests for "Mountain map".
+- Tests for "Create the longest sequence of 1s".
 
 ## [1.6.0] - 2022-03-11
 ### Added